排队叫号

+++ reservation/src/main/java/com/jishi/reservation/controller/protocol/OutpatientFeeDocVO.java
@@ -4,6 +4,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -18,7 +19,7 @@ public class OutpatientFeeDocVO {
       @ApiModelProperty("单据类型")
       private int documentType;
       @ApiModelProperty("单据金额")
-      private double documentAmount;
+      private BigDecimal documentAmount;
       @ApiModelProperty("开单时间")
       private Date documentDate;
       @ApiModelProperty("是否结算卡支付，0-否，1-是")
@@ -26,5 +27,5 @@ public class OutpatientFeeDocVO {
       @ApiModelProperty("支付状态，1-已支付，0-未支付")
       private int payStatus;
       @ApiModelProperty("已退金额")
-      private double returnNumber;
+      private BigDecimal returnNumber;
 }
